// src/app/common-library/common-types/common-type.tsx
export type InputType =
  'string'
  | 'string-number'
  | 'number'
  | 'date-time'
  | 'search-select'
  | 'file'
  | 'tree-select'
  | 'nested'
  | 'radio';
export type SearchModel = {
  [T: string]: {
    type: InputType;
    placeholder?: string;
    label: string | ReactElement;
    keyField?: string;
    disabled?: boolean | ((values: any) => boolean);
    selectField?: string;
    onSearch?: (e: any, values?: any) => any;
    onChange?: (value: any, props: { setFieldValue: ((name: string, value: any) => void), values: any }) => any;
    data?: any;
    name?: string;
    showTime?: boolean;
    option?: Array<any>;
  };
};

export type MasterBodyColumns = {
  [T: string]: ColumnDescription
} | ColumnDescription[];

export type ModifyForm = { _header: (string | ((entity: any) => string)) } & Panels;

export type Panels = {
  [T: string]: ModifyPanel | string,
}
export type ModifyPanel = { _title: string, _validationField?: string } & InputGroups;

export type InputGroups = {
  [T: string]: ModifyInputGroup | string
};


export type ModifyInputGroup = InputGroupType & {
  _subTitle: string;
  _className?: string;
  _inputClassName?: string;
};

export type InputGroupType = { [T: string]: _ModifyModelInput | string }
// ...
export type _ModifyModelInput =
  ({ _type: 'object', [S: string]: any })
  | ({ _type: 'image', value?: any, pathField?: string, width?: string | number, height?: string | number } & _CommonProps)
  | ({ _type: 'file', value?: any } & _CommonProps)
  | ({ _type: 'string' | 'string-number' | 'email' | 'date-time' | 'number' | 'boolean' | 'tag' | 'gallery' } & _CommonProps)
  | ({ _type: 'custom', component: (value: any) => ReactElement })
  | ({ _type: 'radio', options: { value: any, label: string }[] | ((...props: any) => { value: any, label: string }[]); } & _CommonProps)
  | ({ _type: 'checkbox' } & _CommonProps)
  | ({ _type: 'search-select', onSearch: GetAllPropsServer<any> | GetAllProps<any>, keyField?: string, selectField?: string } & _CommonProps)
  | ({ _type: 'tree-select', onSearch: any, keyField?: string, selectField?: string } & _CommonProps)
  | ({ _type: 'select', onSearch: any, keyField?: string, selectField?: string } & _CommonProps)
// ...
export type RenderInfoDetail = {
  header?: string;
  className?: string;
  titleClassName?: string;
  dataClassName?: string;
  style?: string;
  data: RenderInfoDetailColumn,
}[]
export type RenderInfoDetailColumn = {
  [T: string]: {
    title?: string;
    formatter?: (value: any | any[], entity?: any) => ReactElement;
    keyField?: string;
  }
}